LOGBROKER-10406 Move kafka lib to sdk
