+++ ydb/core/http_proxy/ut/sqs_topic_ut/ya.make
@@ -15,7 +15,7 @@ PEERDIR(
     ydb/core/http_proxy/ut/datastreams_fixture
     ydb/core/persqueue/ut/common
     ydb/core/testlib/default
-    ydb/library/kafka
+    ydb/public/sdk/cpp/src/library/kafka
     ydb/library/aclib
     ydb/library/persqueue/tests
     ydb/public/sdk/cpp/src/client/discovery
