+++ ydb/core/kafka_proxy/actors/control_plane_common.h
@@ -4,7 +4,7 @@
 
 #include "ydb/core/kafka_proxy/actors/actors.h"
 
-#include <ydb/library/kafka/kafka.h>
+#include <ydb/public/sdk/cpp/src/library/kafka/kafka.h>
 
 #include <ydb/core/kafka_proxy/kafka_events.h>
