+++ ydb/core/kafka_proxy/kafka_messages.h
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include <ydb/library/kafka/kafka_messages_int.h>
+#include <ydb/public/sdk/cpp/src/library/kafka/kafka_messages_int.h>
 #include "kafka_consumer_protocol.h"
 
 namespace NKafka {
