+++ ydb/core/persqueue/pqtablet/batching/ut/ya.make
@@ -8,7 +8,7 @@ PEERDIR(
     library/cpp/streams/zstd
     ydb/core/persqueue/public/write_meta
     ydb/core/protos
-    ydb/library/kafka
+    ydb/public/sdk/cpp/src/library/kafka
 )
 
 END()
