+++ ydb/core/persqueue/pqtablet/batching/ya.make
@@ -11,7 +11,7 @@ PEERDIR(
     ydb/core/persqueue/common
     ydb/core/persqueue/events
     ydb/core/persqueue/public/codecs
-    ydb/library/kafka
+    ydb/public/sdk/cpp/src/library/kafka
     ydb/library/persqueue/counter_time_keeper
     ydb/core/persqueue/public/write_meta
     ydb/core/protos
