+++ ydb/core/persqueue/ut/ut_with_sdk/ya.make
@@ -20,7 +20,7 @@ PEERDIR(
     library/cpp/svnversion
     ydb/core/persqueue/ut/common
     ydb/core/testlib/default
-    ydb/library/kafka
+    ydb/public/sdk/cpp/src/library/kafka
     ydb/public/sdk/cpp/src/client/persqueue_public/ut/ut_utils
     ydb/public/sdk/cpp/src/client/topic/ut/ut_utils
