+++ ydb/library/ya.make
@@ -20,7 +20,6 @@ RECURSE(
     http_proxy
     intersection_tree
     json_index
-    kafka
     keys
     kll_median
     lockfree_bucket
