+++ ydb/public/sdk/cpp/src/client/topic/impl/read_session_impl.ipp
@@ -16,7 +16,7 @@
 
 #include <google/protobuf/util/time_util.h>
 
-#include <ydb/library/kafka/kafka_records.h>
+#include <ydb/public/sdk/cpp/src/library/kafka/kafka_records.h>
 
 #include <library/cpp/containers/disjoint_interval_tree/disjoint_interval_tree.h>
