+++ ydb/public/sdk/cpp/src/library/kafka/kafka_log_impl.h
