+++ ydb/public/sdk/cpp/src/library/kafka/kafka_records.h
