+++ ydb/public/sdk/cpp/src/library/kafka/ut/kafka_records_ut.cpp
@@ -1,5 +1,5 @@
-#include <ydb/library/kafka/kafka_messages_int.h>
-#include <ydb/library/kafka/kafka_records.h>
+#include <ydb/public/sdk/cpp/src/library/kafka/kafka_messages_int.h>
+#include <ydb/public/sdk/cpp/src/library/kafka/kafka_records.h>
 
 #include <library/cpp/testing/unittest/registar.h>
