+++ ydb/public/sdk/cpp/src/library/kafka/ya.make
