+++ ydb/public/sdk/cpp/src/library/ya.make
@@ -3,6 +3,7 @@ RECURSE(
     grpc/client
     issue
     jwt
+    kafka
     operation_id
     persqueue/obfuscate
     persqueue/topic_parser_public
